Separate out the speechplayeritem class, implement a queued prefetch
